CI: Reconfigure events
This being an open source project, we want pull requests from branches
on a forked repo to run actions on *our* fork. So we run on pull request
events, and on pushes to (our) master branch (i.e. merges).

Also, we don't want a push to a branch for which there is an open PR to
trigger *two* events (which was the case previously).